requirements.txt
# === FastAPI Server ===
fastapi==0.110.2
uvicorn==0.29.0

# === OpenAI GPT-4 ===
openai==1.30.1

# === ElevenLabs MCP Agent ===
elevenlabs==1.50.2
websockets==12.0
pydub==0.25.1

# === Vonage Voice API ===
vonage==3.6.9

# === Google Sheets Logging ===
gspread==5.12.4
google-auth==2.29.0
oauth2client==4.1.3

# === Utilities ===
python-dotenv==1.0.1
requests==2.32.3
jinja2==3.1.4
apscheduler==3.10.4
